requirements.txt
flask
flask-restful
flask-pymongo
webargs
simple-crypt
pycrypto
cryptography
PyJWT
